Part6/3_3_A2A_MCP_LangGraph/requirements.txt
requests
uvicorn
fastmcp
starlette
python-dotenv
a2a-sdk
langgraph
langchain-core
langchain-google-vertexai
langchain-mcp-adapters